+++ pkg/binder/plugins/k8s-plugins/k8s_plugins.go
@@ -9,12 +9,9 @@ import (
 	"sync"
 
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apiserver/pkg/util/feature"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes"
 	ksf "k8s.io/kube-scheduler/framework"
-	"k8s.io/kubernetes/pkg/features"
-	k8sframework "k8s.io/kubernetes/pkg/scheduler/framework"
 	k8splfeature "k8s.io/kubernetes/pkg/scheduler/framework/plugins/feature"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
@@ -45,7 +42,7 @@ func New(
 	var k8sPlugins []common.K8sPlugin
 	k8sFramework, k8sFeatures := newK8sPluginDependencies(client, informerFactory)
 
-	for _, initFunc := range []func(k8sframework.Handle, *k8splfeature.Features, int64) (common.K8sPlugin, error){
+	for _, initFunc := range []func(ksf.Handle, *k8splfeature.Features, int64) (common.K8sPlugin, error){
 		volumebinding.NewVolumeBindingPlugin,
 		dynamicresources.NewDynamicResourcesPlugin,
 	} {
@@ -90,7 +87,6 @@ func newK8sPluginDependencies(
 ) (k8sframework.Handle, *k8splfeature.Features) {
 	k8sFramework := k8s_utils.NewFrameworkHandle(client, informerFactory, nil)
 	k8sFeatures := k8splfeature.Features{
-		// EnableVolumeCapacityPriority:    feature.DefaultFeatureGate.Enabled(features.VolumeCapacityPriority),
 		EnableDynamicResourceAllocation: feature.DefaultFeatureGate.Enabled(features.DynamicResourceAllocation),
 	}